M my_esp_now, AirData, Enums: code for ENS&DHT module M main_ens_dht_sleep: add retries if sensors fail to give data M main_hub: control getting current time (NW!) M index.html (hub): improved chart display

+++ lib/Common/src/AirData.h
@@ -5,8 +5,11 @@
 /// @brief
 struct AirData
 {
+    // /// @brief Temperature in Celzius.
+    // int8_t temperature;
+    
     /// @brief Temperature in Celzius.
-    int8_t temperature;
+    float temperature;
     /// @brief Relative humidity.
     uint8_t humidity;
     /// @brief 0-Normal operation, 1-Warm-Up phase, first 3 minutes after power-on.
